Move lock screen state observation into a separate LifecycleObserver
Use the Engine's lifecycle to trigger listening for lock screen state changes.
